uploadimage/server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic-settings==2.3.4
google-generativeai==0.7.2
pillow==10.4.0
pytesseract==0.3.13
